extract components with options and default values
